Set up generation of API Reference Docs using DocFx (#83)
### Summary

Set up reference doc generation using DocFx.

### Related work items

* https://github.com/temporalio/sdk-dotnet/issues/15
* https://github.com/temporalio/sdk-dotnet/issues/17

#### Note

This PR sets up an automated build of ARI reference docs using DocFx.
We will include it int the CI later.
This change does not affect the actual product code. I do not think that this will benefit from a review at present, so I will proceed with merging it immediately. We will, certainly, need to review and adjust the visuals later. For now it uses the default template.

### Commits

* Set up scaffolding for DocFx-based docs generation.

* DocFx home page renders correctly.

* Working on allowing multiple target docs exist SxS.

* Split DocFxGenerator projects. Prep step works.

* Fixed some warings when generating docs.

* Updated build of `Api.Client.ManagedGrpc` to use commited files.

* Updated Proto-based Stubs.

* One target's docs build well.

* Docs correctly generated.

* Search works

* Moved cleanup back to separate project.

* Massaging ToC

* Everything works.

* Swap `Doc/` and `Docs`.

* Swap all `/Doc/` and `/Docs/` references in all files and references.