Replace Nuget-based use of DocFx with a custom installation and invocation script (#84)
### Summary

Simplify API Ref Doc generation in the build system

### Related work items

* https://github.com/temporalio/sdk-dotnet/issues/15
* https://github.com/temporalio/sdk-dotnet/issues/17

### Details

We recently added API Ref generation using DocFx. Initially, we used DocFx via Nuget.
However, that is flaky and error prone in a multi-targeting context. It was randomly breaking "some" of the time in the build system, and was very complicated.

This change replaces it all with a custom batch script. The script downloads DocFx executables if required and generated Ref Docs for all targets. We invoke the script from the build system. The build system also places the necessary files into the build directories and cleans up after DocFx. Overall, this is both, simpler and more robust.

Currently, the docs are only generated under Windows.

#### Note
 
This change does not affect the actual product code. It only touches the API Ref Docs generation script. I do not think that this will benefit from a review at present, so I will proceed with merging it immediately.